OBGM-47 Refactor ProductPackageExcelImporter
